balabit.logstore: Record flag checker convenience macros.
Implement a set of macros that check whether a given flag is set on a
LogStore record. All functions can work either with an explicitly
specified record, or with *logstore-record*, if bound.

They're also generated by another macro.

For this to work, logstore-record/flag-set? was also implemented,
which does the heavy lifting under the macros.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>