b.l.c.record: Implement decompression.
Pass the data part of the chunk through a series of functions, which -
depending on meta-data - will do the right thing.

For now, we can only decompress data, when the :compressed flag is
set. This is done by transforming the buffer into an InputStream, and
wrapping it in an InflateInputStream. The result can then be slurped
up, or processed further.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>