b.l.core.*: Code cleanup.
Removed the :header member of LSTFileHeader, it's irrelevant for
everyone but the internal reader. Also renamed a couple of other
fields to nicer or more appropriate, less confusing names.

This does go against the LogStore file format spec, because that uses
'record' and 'chunk' in two contexts, each meaning something entirely
different in the other.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>