scripts: Add a script to build docs.
We build two kinds of documentation: the end-user, public-api docs
that only includes the balabit.logstore and balabit.logstore.examples
namespaces, and a developer-oriented one, that includes the rest too.

For maximum satisfaction, the source files are given to marginalia in
a specific order, that makes sense from a documentation point of view.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>