b.l.c.r.chunk: Convert unserialized string splitting to use gloss.
Instead of parsing and splitting the buffer containing unserialized
strings, use gloss for the job. This requires that the output of
decompression be available as a ByteBuffer, so that part was changed
too.

Also, message splitting and deserialization was baked into a single
multi-method, as it makes little sense to have them separate. This
part might be reverted in the future, depends on how deserialization
will work in the end.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>