Add Clojure 1.5 profile and a with-all-profiles alias.
Just for the heck of it, test with clojure 1.5 snapshot too (and add
Sonatype repositories for that to work).

And for future comptibility, add a with-all-profiles alias to
project.clj, so the travis.yml file won't have to be modified if more
profiles get added in the future: all interesting profiles will be
controlled from a single place now, the project.clj file.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>