Use a separate profile for building the docs
There is no need whatsoever to include marginalia and lein-marg in the
base dependencies - it's completely unnecessary for normal operation,
it's only needed to build the docs.

For this reason, this patch creates a new profile, 'doc', that has a
marginalia dependency. Furthermore, scripts/build-docs.clj was turned
into a namespace, and the lein-exec dependency was dropped.

We can now do:
 lein2 with-profile doc run -m balabit.logstore.doc/generate-docs

And that will build the docs. And the 'build-docs' alias does just
that, too.

The travis script was updated to test build-docs too.

While there, also moved the lein-midje plugin into the 'dev' profile,
so we have no plugins by default.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>