A complete rewrite
This is a complete rewrite of the library. It is more featureful, more
extensible, and soon will be better documented too. There is very
little left from the original design, the new one is much easier to
work with, as the library was turned from an object-oriented approach
into something far more data-driven.

The end result is that LogStores are now represented as lazily loaded
maps, and one can work with the results using idiomatic Clojure
functions.

The library was also relicensed, from GPLv3 to CC-BY-SA 3.0, for
compatibility with the EPL.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>