java: Reworked, more Java-like API
The API was reworked a little, to provide a LogStore and a LogStoreMap
class. The latter is a wrapper around normal clojure maps, that looks
up the keys by converting them to keywords first.

The latter encapsulates the parsing into its constructor, and
otherwise works as a map.

In the future, LogStoreMap will be extended to implement the
java.util.Map interface.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>