bin/lgstool: Small wrapper around the CLI
The main use of the wrapper is that one can put it on the PATH, and
use a git checkout of the library & tool simply. It will find the repo
(assumes it is in the repo, and symlinked or directly put on PATH, not
copied), and run lein lgstool, with any relative filenames fully
resolved.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>